Code fixes, most of common_types removed, all tests passing and program runs

+++ pkg/population_factory/population_factory.go
@@ -6,7 +6,7 @@ import (
 
 // IPopulationFactory interface for population factories
 type IPopulationFactory interface {
-	CreatePopulation(populationSize int) (map[int]i_agent.IAgent, error)
+	CreatePopulation(populationSize int) (map[int64]i_agent.IAgent, error)
 }
 
 // BasicPopulationFactroy is a basic variant of IPopulationFactory
@@ -20,8 +20,8 @@ func NewBasicPopulationFactroy() *BasicPopulationFactroy {
 }
 
 // CreatePopulation is used to creating initial population
-func (b *BasicPopulationFactroy) CreatePopulation(populationSize int) (map[int]i_agent.IAgent, error) {
-	var population = make(map[int]i_agent.IAgent)
+func (b *BasicPopulationFactroy) CreatePopulation(populationSize int) (map[int64]i_agent.IAgent, error) {
+	var population = make(map[int64]i_agent.IAgent)
 	for i := 0; i < populationSize; i++ {
 		// TODO
 		// population[i] = *agent.NewAgent(i)
